fast/requirements.txt
fastapi
uvicorn
pyjwt
pydantic
sqlalchemy
alembic
python-dotenv
pymysql
mysqlclient
httpx
faker
python-multipart
pytest
